refactor: move SlotContextMenu to src/components/graph/
Colocate with NodeContextMenu since both are siblings in the
GraphCanvas template tree.

Addresses review feedback:
https://github.com/Comfy-Org/ComfyUI_frontend/pull/9045#discussion_r2925519408
