+++ src/components/graph/GraphCanvas.vue
@@ -155,7 +155,7 @@ import NodeContextMenu from '@/components/graph/NodeContextMenu.vue'
 import NodeDragPreview from '@/components/graph/NodeDragPreview.vue'
 import NodeSelectionModeBanner from '@/components/graph/NodeSelectionModeBanner.vue'
 import SelectionToolbox from '@/components/graph/SelectionToolbox.vue'
-import SlotContextMenu from '@/renderer/extensions/vueNodes/components/SlotContextMenu.vue'
+import SlotContextMenu from '@/components/graph/SlotContextMenu.vue'
 import TitleEditor from '@/components/graph/TitleEditor.vue'
 import NodePropertiesPanel from '@/components/rightSidePanel/RightSidePanel.vue'
 import { useAgentDockMount } from '@/workbench/extensions/agent/composables/useAgentDockMount'
