+++ src/components/graph/SlotContextMenu.vue
